new features: 1. administrators can now only get goods/coupon created by themselves.
bugs fixed:
1. if the result of System.getProperty("user.dir") is "/", it will be removed, or throwing exception in linux system

+++ src/main/java/com/tsong/cmall/entity/GoodsInfo.java
@@ -40,12 +40,12 @@ public class GoodsInfo {
 
     private Byte goodsSaleStatus;
 
-    private Integer createUser;
+    private Long createUser;
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
     private Date createTime;
 
-    private Integer updateUser;
+    private Long updateUser;
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
     private Date updateTime;
